+++ app/src/main/java/com/forrestguice/suntimeswidget/map/WorldMapView.java
@@ -1,5 +1,5 @@
 /**
-    Copyright (C) 2018 Forrest Guice
+    Copyright (C) 2018-2019 Forrest Guice
     This file is part of SuntimesWidget.
 
     SuntimesWidget is free software: you can redistribute it and/or modify
@@ -18,19 +18,27 @@
 package com.forrestguice.suntimeswidget.map;
 
 import android.annotation.SuppressLint;
+import android.app.ProgressDialog;
 import android.content.Context;
+import android.content.Intent;
 import android.content.res.TypedArray;
 import android.graphics.Bitmap;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v4.content.ContextCompat;
+import android.support.v4.content.FileProvider;
 import android.util.AttributeSet;
 import android.util.Log;
+import android.widget.Toast;
 
+import com.forrestguice.suntimeswidget.ExportTask;
 import com.forrestguice.suntimeswidget.R;
 import com.forrestguice.suntimeswidget.calculator.SuntimesRiseSetDataset;
 import com.forrestguice.suntimeswidget.themes.SuntimesTheme;
+
+import java.io.File;
 
 public class WorldMapView extends android.support.v7.widget.AppCompatImageView
 {
@@ -68,8 +76,7 @@ public class WorldMapView extends android.support.v7.widget.AppCompatImageView
         if (isInEditMode())
         {
             setBackgroundColor(Color.WHITE);
-            Bitmap b = Bitmap.createBitmap(256, 256, Bitmap.Config.ARGB_8888);
-            setImageBitmap(b);
+            setImageBitmap(Bitmap.createBitmap(256, 256, Bitmap.Config.ARGB_8888));
         }
         setMapMode(context, mode);
         themeViews(context);
@@ -338,6 +345,105 @@ public class WorldMapView extends android.support.v7.widget.AppCompatImageView
     public static abstract class WorldMapTaskListener
     {
         public void onFinished( Bitmap result ) {}
+    }
+
+
+    @Override
+    public void setImageBitmap(Bitmap b)
+    {
+        this.bitmap = b;
+        super.setImageBitmap(b);
+    }
+
+    private Bitmap bitmap;
+    public Bitmap getBitmap()
+    {
+        return bitmap;
+    }
+
+    public void shareBitmap()
+    {
+        if (bitmap != null)
+        {
+            WorldMapExportTask exportTask = new WorldMapExportTask(getContext(), "SuntimesWorldMap", true, true);
+            exportTask.setTaskListener(new ExportTask.TaskListener()
+            {
+                @Override
+                public void onStarted()
+                {
+                    showProgress();
+                }
+
+                @Override
+                public void onFinished(ExportTask.ExportResult result)
+                {
+                    dismissProgress();
+
+                    Context context = getContext();
+                    if (context != null)
+                    {
+                        if (result.getResult())
+                        {
+                            Intent shareIntent = new Intent();
+                            shareIntent.setAction(Intent.ACTION_SEND);
+                            shareIntent.setType(result.getMimeType());
+                            shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+
+                            try {
+                                Uri shareURI = FileProvider.getUriForFile(context, "com.forrestguice.suntimeswidget.fileprovider", result.getExportFile());
+                                shareIntent.putExtra(Intent.EXTRA_STREAM, shareURI);
+
+                                String successMessage = context.getString(R.string.msg_export_success, result.getExportFile().getAbsolutePath());
+                                Toast.makeText(context.getApplicationContext(), successMessage, Toast.LENGTH_LONG).show();
+
+                                context.startActivity(Intent.createChooser(shareIntent, context.getResources().getText(R.string.msg_export_to)));
+                                return;   // successful export ends here...
+
+                            } catch (Exception e) {
+                                Log.e(LOGTAG, "Failed to share file URI! " + e);
+                            }
+
+                        } else {
+                            File file = result.getExportFile();
+                            String path = ((file != null) ? file.getAbsolutePath() : "<path>");
+                            Toast.makeText(context.getApplicationContext(), context.getString(R.string.msg_export_failure, path), Toast.LENGTH_LONG).show();
+                        }
+                    }
+                }
+            });
+            exportTask.setBitmaps(new Bitmap[] { bitmap });
+            exportTask.execute();
+
+        } else Log.w(LOGTAG, "shareBitmap: null!");
+    }
+
+    private ProgressDialog progressDialog;
+    private void showProgress()
+    {
+        dismissProgress();
+        Context context = getContext();
+        if (context != null)
+        {
+            progressDialog = new ProgressDialog(context);
+            progressDialog.show();
+        }
+    }
+    private void dismissProgress()
+    {
+        if (progressDialog != null)
+        {
+            if (progressDialog.isShowing()) {
+                progressDialog.dismiss();
+            }
+            progressDialog = null;
+        }
+    }
+
+    @Override
+    public void onDetachedFromWindow()
+    {
+        super.onDetachedFromWindow();
+        dismissProgress();
     }
 
 }