+++ app/src/main/java/com/forrestguice/suntimeswidget/themes/SuntimesTheme.java
@@ -1,5 +1,5 @@
 /**
-    Copyright (C) 2014-2018 Forrest Guice
+    Copyright (C) 2014-2019 Forrest Guice
     This file is part of SuntimesWidget.
 
     SuntimesWidget is free software: you can redistribute it and/or modify
@@ -50,6 +50,7 @@ public class SuntimesTheme
     public static final String THEME_TIMECOLOR = "timecolor";
     public static final String THEME_TIMESUFFIXCOLOR = "timesuffixcolor";
     public static final String THEME_ACTIONCOLOR = "actioncolor";
+    public static final String THEME_ACCENTCOLOR = "accentcolor";
 
     public static final String THEME_SUNRISECOLOR = "sunrisecolor";
     public static final String THEME_NOONCOLOR = "nooncolor";
@@ -144,6 +145,7 @@ public class SuntimesTheme
     protected int themeTimeColor;
     protected int themeTimeSuffixColor;
     protected int themeActionColor;
+    protected int themeAccentColor;
 
     protected int themeSunriseTextColor;
     protected int themeSunriseIconColor;
@@ -316,6 +318,7 @@ public class SuntimesTheme
         this.themeTimeColor = themes.getInt( theme + THEME_TIMECOLOR, defaultTheme.themeTimeColor );
         this.themeTimeSuffixColor = themes.getInt( theme + THEME_TIMESUFFIXCOLOR, defaultTheme.themeTimeSuffixColor );
         this.themeActionColor = themes.getInt( theme + THEME_ACTIONCOLOR, defaultTheme.themeActionColor );
+        this.themeAccentColor = themes.getInt( theme + THEME_ACCENTCOLOR, defaultTheme.themeAccentColor );
 
         this.themeSunriseTextColor = themes.getInt( theme + THEME_SUNRISECOLOR, defaultTheme.themeSunriseTextColor );
         this.themeSunriseIconColor = themes.getInt( theme + THEME_RISEICON_FILL_COLOR, defaultTheme.themeSunriseIconColor );
@@ -399,6 +402,7 @@ public class SuntimesTheme
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_TIMECOLOR, this.themeTimeColor);
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_TIMESUFFIXCOLOR, this.themeTimeSuffixColor);
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_ACTIONCOLOR, this.themeActionColor);
+        themePrefs.putInt(themePrefix + SuntimesTheme.THEME_ACCENTCOLOR, this.themeAccentColor);
 
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_SUNRISECOLOR, this.themeSunriseTextColor);
         themePrefs.putInt(themePrefix + SuntimesTheme.THEME_RISEICON_FILL_COLOR, this.themeSunriseIconColor);
@@ -488,6 +492,7 @@ public class SuntimesTheme
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_TIMECOLOR);
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_TIMESUFFIXCOLOR);
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_ACTIONCOLOR);
+        themePrefs.remove(themePrefix + SuntimesTheme.THEME_ACCENTCOLOR);
 
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_SUNRISECOLOR);
         themePrefs.remove(themePrefix + SuntimesTheme.THEME_RISEICON_FILL_COLOR);
@@ -853,6 +858,11 @@ public class SuntimesTheme
     public int getActionColor()
     {
         return themeActionColor;
+    }
+
+    public int getAccentColor()
+    {
+        return themeAccentColor;
     }
 
     public int[] getPadding()