+++ app/src/main/java/com/forrestguice/suntimeswidget/themes/SuntimesThemeXML.java
@@ -1,5 +1,5 @@
 /**
-    Copyright (C) 2017-2018 Forrest Guice
+    Copyright (C) 2017-2019 Forrest Guice
     This file is part of SuntimesWidget.
 
     SuntimesWidget is free software: you can redistribute it and/or modify
@@ -31,6 +31,7 @@ import org.xmlpull.v1.XmlSerializer;
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.util.ArrayList;
 
 @SuppressWarnings("Convert2Diamond")
@@ -38,7 +39,7 @@ public class SuntimesThemeXML implements SuntimesThemeIO
 {
     public static final String KEY_THEMES = "themes";
     public static final String KEY_THEME = "theme";
-    public static final String VERSION = "1.4";
+    public static final String VERSION = "1.5";
 
     public SuntimesThemeXML() {}
 
@@ -148,7 +149,7 @@ public class SuntimesThemeXML implements SuntimesThemeIO
      * @throws IOException if failed to write to out
      */
     @Override
-    public boolean write(Context context, BufferedOutputStream out, SuntimesTheme[] themes) throws IOException
+    public boolean write(Context context, OutputStream out, SuntimesTheme[] themes) throws IOException
     {
         signalExportStarted();
         if (themes != null)
@@ -214,6 +215,10 @@ public class SuntimesThemeXML implements SuntimesThemeIO
                 xml.startTag(namespace, SuntimesTheme.THEME_ACTIONCOLOR);
                 xml.text(colorToString(theme.getActionColor()));
                 xml.endTag(namespace, SuntimesTheme.THEME_ACTIONCOLOR);
+
+                xml.startTag(namespace, SuntimesTheme.THEME_ACCENTCOLOR);
+                xml.text(colorToString(theme.getAccentColor()));
+                xml.endTag(namespace, SuntimesTheme.THEME_ACCENTCOLOR);
 
                 xml.startTag(namespace, SuntimesTheme.THEME_SUNRISECOLOR);
                 xml.text(colorToString(theme.getSunriseTextColor()));
@@ -534,6 +539,8 @@ public class SuntimesThemeXML implements SuntimesThemeIO
                         theme.themeTimeSuffixColor = colorStringToInt(value);
                     } else if (tag.equalsIgnoreCase(SuntimesTheme.THEME_ACTIONCOLOR)) {
                         theme.themeActionColor = colorStringToInt(value);
+                    } else if (tag.equalsIgnoreCase(SuntimesTheme.THEME_ACCENTCOLOR)) {
+                        theme.themeAccentColor = colorStringToInt(value);
 
                     } else if (tag.equalsIgnoreCase(SuntimesTheme.THEME_SUNRISECOLOR)) {
                         theme.themeSunriseTextColor = colorStringToInt(value);
@@ -640,6 +647,9 @@ public class SuntimesThemeXML implements SuntimesThemeIO
         }
         if (theme.themeActionColor == 0) {
             theme.themeActionColor = theme.themeSunsetIconColor;
+        }
+        if (theme.themeAccentColor == 0) {
+            theme.themeAccentColor = theme.themeSunsetIconColor;
         }
         return theme;
     }